requirements.txt
flask
flask_socketio
scikit-learn
eventlet
python-dotenv